+++ packages/orq-rc/src/orq_ai_sdk/models/createagentrequest_responsebody.py
@@ -1558,7 +1558,6 @@ class CreateAgentRequestResponseBodyTypedDict(TypedDict):
     id: str
     key: str
     r"""Unique identifier for the agent within the workspace"""
-    display_name: str
     project_id: str
     role: str
     description: str
@@ -1577,6 +1576,7 @@ class CreateAgentRequestResponseBodyTypedDict(TypedDict):
     r"""Array of memory store identifiers. Accepts both memory store IDs and keys."""
     team_of_agents: List[CreateAgentRequestTeamOfAgentsTypedDict]
     r"""The agents that are accessible to this orchestrator. The main agent can hand off to these agents to perform tasks."""
+    display_name: NotRequired[str]
     created_by_id: NotRequired[Nullable[str]]
     updated_by_id: NotRequired[Nullable[str]]
     created: NotRequired[str]
@@ -1599,8 +1599,6 @@ class CreateAgentRequestResponseBody(BaseModel):
 
     key: str
     r"""Unique identifier for the agent within the workspace"""
-
-    display_name: str
 
     project_id: str
 
@@ -1628,6 +1626,8 @@ class CreateAgentRequestResponseBody(BaseModel):
 
     team_of_agents: List[CreateAgentRequestTeamOfAgents]
     r"""The agents that are accessible to this orchestrator. The main agent can hand off to these agents to perform tasks."""
+
+    display_name: Optional[str] = None
 
     created_by_id: OptionalNullable[str] = UNSET
 
@@ -1657,6 +1657,7 @@ class CreateAgentRequestResponseBody(BaseModel):
     def serialize_model(self, handler):
         optional_fields = set(
             [
+                "display_name",
                 "created_by_id",
                 "updated_by_id",
                 "created",