+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ60BRVD6QBCBS98PA5H5EXB"
+        "01KJ7FDGHVBY30N8ZM9CYF2J5A"
     )
     r"""The unique identifier of the data source"""