+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveagentrequestop.py
@@ -2135,7 +2135,6 @@ class RetrieveAgentRequestResponseBodyTypedDict(TypedDict):
     id: str
     key: str
     r"""Unique identifier for the agent within the workspace"""
-    display_name: str
     workspace_id: str
     project_id: str
     role: str
@@ -2155,6 +2154,7 @@ class RetrieveAgentRequestResponseBodyTypedDict(TypedDict):
     r"""Array of memory store identifiers. Accepts both memory store IDs and keys."""
     team_of_agents: List[RetrieveAgentRequestTeamOfAgentsTypedDict]
     r"""The agents that are accessible to this orchestrator. The main agent can hand off to these agents to perform tasks."""
+    display_name: NotRequired[str]
     created_by_id: NotRequired[Nullable[str]]
     updated_by_id: NotRequired[Nullable[str]]
     created: NotRequired[str]
@@ -2177,8 +2177,6 @@ class RetrieveAgentRequestResponseBody(BaseModel):
 
     key: str
     r"""Unique identifier for the agent within the workspace"""
-
-    display_name: str
 
     workspace_id: str
 
@@ -2209,6 +2207,8 @@ class RetrieveAgentRequestResponseBody(BaseModel):
     team_of_agents: List[RetrieveAgentRequestTeamOfAgents]
     r"""The agents that are accessible to this orchestrator. The main agent can hand off to these agents to perform tasks."""
 
+    display_name: Optional[str] = None
+
     created_by_id: OptionalNullable[str] = UNSET
 
     updated_by_id: OptionalNullable[str] = UNSET
@@ -2237,6 +2237,7 @@ class RetrieveAgentRequestResponseBody(BaseModel):
     def serialize_model(self, handler):
         optional_fields = set(
             [
+                "display_name",
                 "created_by_id",
                 "updated_by_id",
                 "created",