+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -2193,7 +2193,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KJ60BR85Z4N1Q1Q1H56XV20A"
+    id: Optional[str] = "01KJ7FDG2T0S9VA4ZJD57PJYM7"
 
     description: Optional[str] = None