+++ packages/orq-rc/RELEASES.md
@@ -6049,3 +6049,13 @@ Based on:
 - [python v4.4.0-rc.37] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.37] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.37 - packages/orq-rc
+
+## 2026-02-24 09:26:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.725.0 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.0] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.0 - packages/orq-rc